create all policy

+++ app/Http/Controllers/BannerController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Banner;
+use App\User;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -16,10 +17,15 @@ class BannerController extends Controller
      */
     public function index()
     {
-        $banners = Banner::latest()->paginate(10);
-        return view ('admin.banner.index', [
-            'banners' => $banners,
-        ]);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('viewAny', Banner::class) ) {
+            $banners = Banner::latest()->paginate(10);
+            return view ('admin.banner.index', [
+                'banners' => $banners,
+            ]);
+        } else {
+            return view ('admin.errors.auth');
+        }
     }
 
     /**
@@ -29,7 +35,12 @@ class BannerController extends Controller
      */
     public function create()
     {
-        return view ('admin.banner.create');
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('viewAny', Banner::class) ) {
+            return view ('admin.banner.create');
+        } else {
+            return view ('admin.errors.auth');
+        }
     }
 
     /**
@@ -40,65 +51,71 @@ class BannerController extends Controller
      */
     public function store(Request $request)
     {
-        $request['trueTitle'] = $request->input('title');
-        $request['title'] = Str::slug($request->input('title'));
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('create', Banner::class) ) {
+            $request['trueTitle'] = $request->input('title');
+            $request['title'] = Str::slug($request->input('title'));
 
-        $validator = Validator::make($request->all(), [
-            'title' => 'required|unique:banners,slug|max:255',
-            'image' => 'required|mimes:jpeg,png,jpg,gif,svg,webp',
-            'url' => 'nullable|url',
-            'target' => 'required',
-            'position' =>  'required|integer|min:0',
-            'is_active' => 'integer|boolean',
-        ], [
-            'title.required' => 'Tên không được để trống',
-            'title.unique' => 'Tên bị trùng',
-            'image.required' => 'Ảnh không được để trống',
-            'image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
-            'url.url' => 'Dữ liệu không đúng định dạng',
-            'target.required' => 'Yêu cầu không được để trống',
-            'position.integer' => 'Sai kiểu dữ liệu',
-            'position.min' => '',
-            'is_active.integer' => 'Sai kiểu dữ liệu',
-            'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
+            $validator = Validator::make($request->all(), [
+                'title' => 'required|unique:banners,slug|max:255',
+                'image' => 'required|mimes:jpeg,png,jpg,gif,svg,webp',
+                'url' => 'nullable|url',
+                'target' => 'required',
+                'position' =>  'required|integer|min:0',
+                'is_active' => 'integer|boolean',
+            ], [
+                'title.required' => 'Tên không được để trống',
+                'title.unique' => 'Tên bị trùng',
+                'image.required' => 'Ảnh không được để trống',
+                'image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
+                'url.url' => 'Dữ liệu không đúng định dạng',
+                'target.required' => 'Yêu cầu không được để trống',
+                'position.integer' => 'Sai kiểu dữ liệu',
+                'position.min' => '',
+                'is_active.integer' => 'Sai kiểu dữ liệu',
+                'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
 
-        ]);
+            ]);
 
-        $errs = $validator->errors();
+            $errs = $validator->errors();
 
-        if ( $validator->fails() ) {
-            return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
-        } else {
-            $banner = new Banner;
-            $banner->title = $request->input('trueTitle');
-            $banner->slug = $request->input('title');
-
-            if ($request->hasFile('image')) {
-                // get file
-                $file = $request->file('image');
-                // get ten
-                $filename = time().'_'.$file->getClientOriginalName();
-                // duong dan upload
-                $path_upload = 'uploads/banner/';
-                // upload file
-    
-                $banner->image = $path_upload.$filename;
-            }
-
-            $banner->url = $request->input('url');
-            $banner->target = $request->input('target');
-            $banner->position = (int)$request->input('position');
-            $banner->is_active = (int)$request->input('is_active');
-
-            if ($banner->save()) {
-                // upload file
-                $request->file('image')->move($path_upload,$filename);
-
-                return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
-
+            if ( $validator->fails() ) {
+                return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
             } else {
-                return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
+                $banner = new Banner;
+                $banner->title = $request->input('trueTitle');
+                $banner->slug = $request->input('title');
+
+                if ($request->hasFile('image')) {
+                    // get file
+                    $file = $request->file('image');
+                    // get ten
+                    $filename = time().'_'.$file->getClientOriginalName();
+                    // duong dan upload
+                    $path_upload = 'uploads/banner/';
+                    // upload file
+        
+                    $banner->image = $path_upload.$filename;
+                }
+
+                $banner->url = $request->input('url');
+                $banner->target = $request->input('target');
+                $banner->position = (int)$request->input('position');
+                $banner->is_active = (int)$request->input('is_active');
+
+                if ($banner->save()) {
+                    // upload file
+                    $request->file('image')->move($path_upload,$filename);
+
+                    return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
+
+                } else {
+                    return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
+                }
             }
+        } else {
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+
         }
     }
 
@@ -121,10 +138,15 @@ class BannerController extends Controller
      */
     public function edit($id)
     {
-        $banner = Banner::findOrFail($id);
-        return view ('admin.banner.edit', [
-            'banner' => $banner,
-        ]);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('update', Banner::class) ) {
+            $banner = Banner::findOrFail($id);
+            return view ('admin.banner.edit', [
+                'banner' => $banner,
+            ]);
+        } else {
+            return view ('admin.errors.auth');
+        }  
     }
 
     /**
@@ -136,72 +158,80 @@ class BannerController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $banner = Banner::find($id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('update', Banner::class) ) {
 
-        if (empty($banner)) {
-            return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
-        }
-
-        $request['trueTitle'] = $request->input('title');
-        $request['title'] = Str::slug($request->input('title'));
-
-        $validator = Validator::make($request->all(), [
-            'title' => 'required|max:255|unique:banners,slug,'.$id,
-            'new_image' => 'nullable|mimes:jpeg,png,jpg,gif,svg,webp',
-            'url' => 'nullable|url',
-            'target' => 'required',
-            'position' =>  'required|integer|min:0',
-            'is_active' => 'integer|boolean',
-        ], [
-            'title.required' => 'Tên không được để trống',
-            'title.unique' => 'Tên bị trùng',
-            'new_image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
-            'url.url' => 'Dữ liệu không đúng định dạng',
-            'target.required' => 'Yêu cầu không được để trống',
-            'position.integer' => 'Sai kiểu dữ liệu',
-            'position.min' => 'Giá trị yêu cầu lớn hơn 0',
-            'is_active.integer' => 'Sai kiểu dữ liệu',
-            'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
-
-        ]);
-
-        $errs = $validator->errors();
-
-        if ( $validator->fails() ) {
-            return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
-        } else {
-            $banner->title = $request->input('trueTitle');
-            $banner->slug = $request->input('title');
-
-            if ($request->hasFile('new_image')) {
-                // xóa file cũ
-                @unlink(public_path($banner->image));
-                // get file mới
-                $file = $request->file('new_image');
-                // get tên
-                $filename = time().'_'.$file->getClientOriginalName();
-                // duong dan upload
-                $path_upload = 'uploads/banner/';
-                // upload file
-
-                $banner->image = $path_upload.$filename;
+            $banner = Banner::find($id);
+            if (empty($banner)) {
+                return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
             }
 
-            $banner->url = $request->input('url');
-            $banner->target = $request->input('target');
-            $banner->position = (int)$request->input('position');
-            $banner->is_active = (int)$request->input('is_active');
+            $request['trueTitle'] = $request->input('title');
+            $request['title'] = Str::slug($request->input('title'));
 
-            if ($banner->save()) {
-                // upload file
-                ( $request->hasFile('new_image') ) ? $request->file('new_image')->move($path_upload,$filename) : '';  
+            $validator = Validator::make($request->all(), [
+                'title' => 'required|max:255|unique:banners,slug,'.$id,
+                'new_image' => 'nullable|mimes:jpeg,png,jpg,gif,svg,webp',
+                'url' => 'nullable|url',
+                'target' => 'required',
+                'position' =>  'required|integer|min:0',
+                'is_active' => 'integer|boolean',
+            ], [
+                'title.required' => 'Tên không được để trống',
+                'title.unique' => 'Tên bị trùng',
+                'new_image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
+                'url.url' => 'Dữ liệu không đúng định dạng',
+                'target.required' => 'Yêu cầu không được để trống',
+                'position.integer' => 'Sai kiểu dữ liệu',
+                'position.min' => 'Giá trị yêu cầu lớn hơn 0',
+                'is_active.integer' => 'Sai kiểu dữ liệu',
+                'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
 
-                return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
+            ]);
 
+            $errs = $validator->errors();
+
+            if ( $validator->fails() ) {
+                return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
             } else {
-                return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
+                $banner->title = $request->input('trueTitle');
+                $banner->slug = $request->input('title');
+
+                if ($request->hasFile('new_image')) {
+                    // xóa file cũ
+                    @unlink(public_path($banner->image));
+                    // get file mới
+                    $file = $request->file('new_image');
+                    // get tên
+                    $filename = time().'_'.$file->getClientOriginalName();
+                    // duong dan upload
+                    $path_upload = 'uploads/banner/';
+                    // upload file
+
+                    $banner->image = $path_upload.$filename;
+                }
+
+                $banner->url = $request->input('url');
+                $banner->target = $request->input('target');
+                $banner->position = (int)$request->input('position');
+                $banner->is_active = (int)$request->input('is_active');
+
+                if ($banner->save()) {
+                    // upload file
+                    ( $request->hasFile('new_image') ) ? $request->file('new_image')->move($path_upload,$filename) : '';  
+
+                    return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
+
+                } else {
+                    return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
+                }
             }
-        }
+        } else {
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+
+        }  
+
+        
     }
 
     /**
@@ -212,16 +242,22 @@ class BannerController extends Controller
      */
     public function destroy($id)
     {
-        $banner = Banner::find($id);
-        if ( empty($banner) ) {
-            return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
-        }
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('destroy', Banner::class) ) {
+            $banner = Banner::find($id);
+            if ( empty($banner) ) {
+                return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
+            }
 
-        if( Banner::destroy($id) != 0 ) {
-            return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            if( Banner::destroy($id) != 0 ) {
+                return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            } else {
+                return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+
+            }
         } else {
-            return response()->json(['mess' => 'Xóa bản không thành công'], 400);
-
-        }
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+        }  
+        
     }
 }
