+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Category;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Validator;
@@ -16,11 +17,15 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::latest()->paginate(10);
-        // dd($categories);
-        return view ('admin.category.index', [
-            'categories' => $categories,
-        ]);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('viewAny', Category::class) ) {
+            $categories = Category::latest()->paginate(10);
+            return view ('admin.category.index', [
+                'categories' => $categories,
+            ]);
+        } else {
+            return view ('admin.errors.auth');
+        }
     }
 
     /**
@@ -30,7 +35,12 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view ('admin.category.create');
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('create', Category::class) ) {
+            return view ('admin.category.create');
+        } else {
+            return view ('admin.errors.auth');
+        }   
     }
 
     /**
@@ -41,63 +51,70 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {   
-        $request['trueName'] = $request->input('name');
-        $request['name'] = Str::slug($request->input('name'));
+        $currentUser = User::findOrFail(Auth()->user()->id);
 
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|unique:categories,slug|max:255',
-            'image' => 'required|mimes:jpeg,png,jpg,gif,svg,webp',
-            'position' => 'integer', 
-            'is_active' => 'integer|boolean',
-            'is_hot' => 'integer|boolean',
-        ], [
-            'name.required' => 'Tên không được để trống',
-            'name.unique' => 'Tên bị trùng',
-            'image.required' => 'Ảnh không được để trống',
-            'image.image' => 'Ảnh không đúng định dạng',
-            'image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
-            'is_active.integer' => 'Sai kiểu dữ liệu',
-            'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
-            'is_hot.integer' => 'Sai kiểu dữ liệu',
-            'is_hot.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
-            'position.integer' => 'Sai kiểu dữ liệu',
-        ]);
+        if ( $currentUser->can('create', Category::class) ) {
+            $request['trueName'] = $request->input('name');
+            $request['name'] = Str::slug($request->input('name'));
 
-        $errs = $validator->errors();
+            $validator = Validator::make($request->all(), [
+                'name' => 'required|unique:categories,slug|max:255',
+                'image' => 'required|mimes:jpeg,png,jpg,gif,svg,webp',
+                'position' => 'integer', 
+                'is_active' => 'integer|boolean',
+                'is_hot' => 'integer|boolean',
+            ], [
+                'name.required' => 'Tên không được để trống',
+                'name.unique' => 'Tên bị trùng',
+                'image.required' => 'Ảnh không được để trống',
+                'image.image' => 'Ảnh không đúng định dạng',
+                'image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
+                'is_active.integer' => 'Sai kiểu dữ liệu',
+                'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
+                'is_hot.integer' => 'Sai kiểu dữ liệu',
+                'is_hot.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
+                'position.integer' => 'Sai kiểu dữ liệu',
+            ]);
 
-        if ( $validator->fails() ) {
-            return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
-        } else {
-            $category = new Category;
-            $category->name = $request->input('trueName');
-            $category->slug = $request->input('name');
+            $errs = $validator->errors();
 
-            if ($request->hasFile('image')) {
-                // get file
-                $file = $request->file('image');
-                // get ten
-                $filename = time().'_'.$file->getClientOriginalName();
-                // duong dan upload
-                $path_upload = 'uploads/category/';
-                // upload file
-    
-                $category->image = $path_upload.$filename;
-            }
-
-            $category->position = (int)$request->input('position');
-            $category->is_active = (int)$request->input('is_active');
-            $category->is_hot = (int)$request->input('is_hot');
-
-            if ($category->save()) {
-                // upload file
-                $request->file('image')->move($path_upload,$filename);
-                return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
-
+            if ( $validator->fails() ) {
+                return response()->json(['errors' => $errs, 'mess' => 'Thêm bản ghi lỗi'], 400);
             } else {
-                return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
-            }
+                $category = new Category;
+                $category->name = $request->input('trueName');
+                $category->slug = $request->input('name');
 
-        }
+                if ($request->hasFile('image')) {
+                    // get file
+                    $file = $request->file('image');
+                    // get ten
+                    $filename = time().'_'.$file->getClientOriginalName();
+                    // duong dan upload
+                    $path_upload = 'uploads/category/';
+                    // upload file
+        
+                    $category->image = $path_upload.$filename;
+                }
+
+                $category->position = (int)$request->input('position');
+                $category->is_active = (int)$request->input('is_active');
+                $category->is_hot = (int)$request->input('is_hot');
+
+                if ($category->save()) {
+                    // upload file
+                    $request->file('image')->move($path_upload,$filename);
+                    return response()->json(['mess' => 'Thêm bản ghi thành công'], 200);
+
+                } else {
+                    return response()->json(['mess' => 'Thêm bản ghi lỗi'], 500);
+                }
+
+            }
+        } else {
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+        }  
+        
     }
 
     /**
@@ -119,13 +136,18 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
-        $category = Category::find($id);
-        if (empty($category)) {
-            dd(false);
-        }
-        return view ('admin.category.edit', [
-            'category' => $category,
-        ]);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('update', Category::class) ) {
+            $category = Category::find($id);
+            if (empty($category)) {
+                dd(false);
+            }
+            return view ('admin.category.edit', [
+                'category' => $category,
+            ]);
+        } else {
+            return view ('admin.errors.auth');
+        }  
     }
 
     /**
@@ -137,70 +159,75 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $category = Category::find($id);
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('update', Category::class) ) {
+            $category = Category::find($id);
 
-        if ( empty($category) ) {
-            return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
-        }
-
-        $request['trueName'] = $request->input('name');
-        $request['name'] = Str::slug($request->input('name'));
-
-        // dd($request->all());
-
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|max:255|unique:categories,slug,'.$id,
-            'new_image' => 'nullable|mimes:jpeg,png,jpg,gif,svg,webp',
-            'position' => 'integer|min:0', 
-            'is_active' => 'integer|boolean',
-            'is_hot' => 'integer|boolean',
-        ], [
-            'name.required' => 'Tên không được để trống',
-            'name.unique' => 'Tên bị trùng',
-            // 'new_image.required' => 'Ảnh không được để trống',
-            'new_image.image' => 'Ảnh không đúng định dạng',
-            'new_image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
-            'is_active.integer' => 'Sai kiểu dữ liệu',
-            'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
-            'is_hot.integer' => 'Sai kiểu dữ liệu',
-            'is_hot.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
-            'position.integer' => 'Sai kiểu dữ liệu',
-            'position.min' => 'Dữ liệu phải lớn hơn 0',
-        ]); 
-
-        $errs = $validator->errors();
-
-        if ( $validator->fails() ) {
-            return response()->json(['errors' => $errs, 'mess' => 'Sửa bản ghi lỗi'], 400);
-        } else {
-            $category->name = $request->input('trueName');
-            $category->slug = $request->input('name');
-            if ($request->hasFile('new_image')) {
-                // xóa file cũ
-                @unlink(public_path($category->image));
-                // get file mới
-                $file = $request->file('new_image');
-                // get tên
-                $filename = time().'_'.$file->getClientOriginalName();
-                // duong dan upload
-                $path_upload = 'uploads/category/';
-                // upload file
-
-                $category->image = $path_upload.$filename;
+            if ( empty($category) ) {
+                return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
             }
-            $category->position = (int)$request->input('position');
-            $category->is_active = (int)$request->input('is_active');
-            $category->is_hot = (int)$request->input('is_hot');
 
-            if ($category->save()) {
-                ( $request->hasFile('new_image') ) ? $request->file('new_image')->move($path_upload,$filename) : '';              
-                return response()->json(['mess' => 'Sửa ghi thành công'], 200);
+            $request['trueName'] = $request->input('name');
+            $request['name'] = Str::slug($request->input('name'));
 
+            // dd($request->all());
+
+            $validator = Validator::make($request->all(), [
+                'name' => 'required|max:255|unique:categories,slug,'.$id,
+                'new_image' => 'nullable|mimes:jpeg,png,jpg,gif,svg,webp',
+                'position' => 'integer|min:0', 
+                'is_active' => 'integer|boolean',
+                'is_hot' => 'integer|boolean',
+            ], [
+                'name.required' => 'Tên không được để trống',
+                'name.unique' => 'Tên bị trùng',
+                // 'new_image.required' => 'Ảnh không được để trống',
+                'new_image.image' => 'Ảnh không đúng định dạng',
+                'new_image.mimes' => 'Ảnh không đúng định dạng, ảnh phải có đuôi jpeg,png,jpg,gif,svg,webp',
+                'is_active.integer' => 'Sai kiểu dữ liệu',
+                'is_active.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
+                'is_hot.integer' => 'Sai kiểu dữ liệu',
+                'is_hot.boolean' => 'Yêu cầu dữ liệu là dạng boolean',
+                'position.integer' => 'Sai kiểu dữ liệu',
+                'position.min' => 'Dữ liệu phải lớn hơn 0',
+            ]); 
+
+            $errs = $validator->errors();
+
+            if ( $validator->fails() ) {
+                return response()->json(['errors' => $errs, 'mess' => 'Sửa bản ghi lỗi'], 400);
             } else {
-                return response()->json(['mess' => 'Sửa bản ghi lỗi'], 500);
-            }
-        }
+                $category->name = $request->input('trueName');
+                $category->slug = $request->input('name');
+                if ($request->hasFile('new_image')) {
+                    // xóa file cũ
+                    @unlink(public_path($category->image));
+                    // get file mới
+                    $file = $request->file('new_image');
+                    // get tên
+                    $filename = time().'_'.$file->getClientOriginalName();
+                    // duong dan upload
+                    $path_upload = 'uploads/category/';
+                    // upload file
 
+                    $category->image = $path_upload.$filename;
+                }
+                $category->position = (int)$request->input('position');
+                $category->is_active = (int)$request->input('is_active');
+                $category->is_hot = (int)$request->input('is_hot');
+
+                if ($category->save()) {
+                    ( $request->hasFile('new_image') ) ? $request->file('new_image')->move($path_upload,$filename) : '';              
+                    return response()->json(['mess' => 'Sửa ghi thành công'], 200);
+
+                } else {
+                    return response()->json(['mess' => 'Sửa bản ghi lỗi'], 500);
+                }
+            }
+        } else {
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+
+        }  
     }
 
     /**
@@ -211,17 +238,21 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        if ( empty($category) ) {
-            return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
-        }
+        $currentUser = User::findOrFail(Auth()->user()->id);
+        if ( $currentUser->can('destroy', Category::class) ) {
+            $category = Category::find($id);
+            if ( empty($category) ) {
+                return response()->json(['mess' => 'Bản ghi không tồn tại'], 400);
+            }
 
-        if( Category::destroy($id) != 0 ) {
-            return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            if( Category::destroy($id) != 0 ) {
+                return response()->json(['mess' => 'Xóa bản ghi thành công'], 200);
+            } else {
+                return response()->json(['mess' => 'Xóa bản không thành công'], 400);
+
+            }
         } else {
-            return response()->json(['mess' => 'Xóa bản không thành công'], 400);
-
-        }
-        
+            return response()->json(['mess' => 'Thêm bản ghi lỗi', 403]);
+        }  
     }
 }
